# backend/docs/swagger.yaml
basePath: /api
definitions:
  dtos.ChairResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  dtos.CreateChairRequest:
    properties:
      description:
        maxLength: 500
        type: string
      location:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 60
        minLength: 2
        type: string
      status:
        enum:
        - ativa
        - inativa
        type: string
    required:
    - location
    - name
    type: object
  dtos.CreateChairResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
  entities.Availability:
    properties:
      chair:
        allOf:
        - $ref: '#/definitions/entities.Chair'
        description: Relacionamentos
      chair_id:
        type: integer
      created_at:
        type: string
      day_of_week:
        description: 0=Domingo, 6=Sábado
        maximum: 6
        minimum: 0
        type: integer
      end_time:
        description: Formato HH:MM
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      start_time:
        description: Formato HH:MM
        type: string
      updated_at:
        type: string
      valid_from:
        type: string
      valid_to:
        type: string
    required:
    - chair_id
    - day_of_week
    - end_time
    - start_time
    type: object
  entities.Booking:
    properties:
      chair:
        $ref: '#/definitions/entities.Chair'
      chair_id:
        type: integer
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: integer
      notes:
        type: string
      start_time:
        type: string
      status:
        enum:
        - agendado
        - confirmado
        - cancelado
        - concluido
        - falta
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/entities.User'
        description: Relacionamentos
      user_id:
        type: integer
    required:
    - chair_id
    - end_time
    - start_time
    - user_id
    type: object
  entities.Chair:
    properties:
      availabilities:
        items:
          $ref: '#/definitions/entities.Availability'
        type: array
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      location:
        maxLength: 100
        minLength: 2
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      status:
        enum:
        - ativa
        - inativa
        type: string
      updated_at:
        type: string
    required:
    - location
    - name
    type: object
  entities.User:
    properties:
      birth_date:
        type: string
      bookings:
        description: Relacionamentos
        items:
          $ref: '#/definitions/entities.Booking'
        type: array
      cpf:
        type: string
      created_at:
        type: string
      email:
        type: string
      function:
        type: string
      gender:
        enum:
        - masculino
        - feminino
        - outro
        type: string
      id:
        type: integer
      last_login:
        type: string
      name:
        maxLength: 60
        minLength: 2
        type: string
      phone:
        maxLength: 20
        minLength: 10
        type: string
      position:
        type: string
      registration:
        type: string
      requested_role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      role:
        enum:
        - usuario
        - atendente
        - admin
        type: string
      sector:
        type: string
      status:
        enum:
        - pendente
        - aprovado
        - reprovado
        type: string
      updated_at:
        type: string
    required:
    - cpf
    - email
    - name
    - phone
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API REST para sistema de agendamento de cadeiras de massagem
  title: Sistema de Agendamento API
  version: "1.0"
paths:
...
  /chairs:
    get:
      consumes:
      - application/json
      description: Retorna uma lista paginada de todas as cadeiras disponíveis
      parameters:
      - default: 10
        description: Limite de resultados por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Número de registros a pular
        in: query
        name: offset
        type: integer
      - description: Filtrar por status da cadeira
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cadeiras
          schema:
            items:
              $ref: '#/definitions/entities.Chair'
            type: array
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar cadeiras
      tags:
      - chairs
    post:
      consumes:
      - application/json
      description: Cria uma nova cadeira de massagem no sistema (apenas admins)
      parameters:
      - description: Dados da cadeira
        in: body
        name: chair
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateChairRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Cadeira criada com sucesso
          schema:
            $ref: '#/definitions/dtos.CreateChairResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Criar cadeira
      tags:
      - chairs
...
  /chairs/{id}:
    delete:
      consumes:
      - application/json
      description: Exclui uma cadeira do sistema (apenas admins)
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cadeira excluída com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Excluir cadeira
      tags:
      - chairs
    get:
      consumes:
      - application/json
      description: Retorna os dados de uma cadeira específica pelo ID
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cadeira encontrada
          schema:
            $ref: '#/definitions/dtos.ChairResponse'
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar cadeira por ID
      tags:
      - chairs
    put:
      consumes:
      - application/json
      description: Atualiza os dados de uma cadeira existente (apenas admins)
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      - description: Dados atualizados da cadeira
        in: body
        name: chair
        required: true
        schema:
          $ref: '#/definitions/entities.Chair'
      produces:
      - application/json
      responses:
        "200":
          description: Cadeira atualizada com sucesso
          schema:
            $ref: '#/definitions/entities.Chair'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Atualizar cadeira
      tags:
      - chairs
  /chairs/{id}/activate:
    post:
      consumes:
      - application/json
      description: Ativa uma cadeira específica (apenas admin)
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cadeira ativada com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Ativar cadeira
      tags:
      - chairs
  /chairs/{id}/deactivate:
    post:
      consumes:
      - application/json
      description: Desativa uma cadeira específica (apenas admin)
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cadeira desativada com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Desativar cadeira
      tags:
      - chairs
  /chairs/{id}/status:
    put:
      consumes:
      - application/json
      description: Altera o status de uma cadeira específica (apenas admin)
      parameters:
      - description: ID da cadeira
        in: path
        name: id
        required: true
        type: integer
      - description: Novo status da cadeira
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status alterado com sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Sem permissão
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Cadeira não encontrada
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Alterar status da cadeira
      tags:
      - chairs
  /chairs/active:
    get:
      consumes:
      - application/json
      description: Lista cadeiras ativas com paginação
      parameters:
      - default: 10
        description: Limite de registros por página
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cadeiras ativas
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parâmetros inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar cadeiras ativas
      tags:
      - chairs
  /chairs/available:
    get:
      consumes:
      - application/json
      description: Lista cadeiras disponíveis para agendamento
      produces:
      - application/json
      responses:
        "200":
          description: Lista de cadeiras disponíveis
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Erro ao buscar cadeiras
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Buscar cadeiras disponíveis
      tags:
      - chairs
  /chairs/stats:
    get:
      consumes:
      - application/json
      description: Retorna estatísticas gerais das cadeiras
      produces:
      - application/json
      responses:
        "200":
          description: Estatísticas das cadeiras
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Erro interno do servidor
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Estatísticas das cadeiras
      tags:
      - chairs
...
securityDefinitions:
  Bearer:
    description: Digite 'Bearer' seguido de um espaço e o token JWT obtido no login
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"